+++ packages/textlint-tester/test/textlint-tester-error-format.ts
@@ -1,10 +1,9 @@
 // LICENSE : MIT
 "use strict";
 import * as assert from "node:assert";
-import { describe, it } from "vitest";
-import { createTestLinter, createTextlintKernelDescriptor } from "../src/textlint-tester.js";
-import { testValid, testInvalid } from "../src/test-util.js";
-import rule from "./fixtures/rule/no-todo.js";
+import { createTestLinter, createTextlintKernelDescriptor } from "../lib/src/textlint-tester.js";
+import { testValid, testInvalid } from "../lib/src/test-util.js";
+import rule from "./fixtures/rule/no-todo";
 
 describe("Error format", () => {
     const textlint = createTestLinter(
