+++ packages/textlint-tester/test/textlint-tester-invalid-testcofig-test.ts
@@ -1,11 +1,10 @@
 import * as assert from "node:assert";
-import { describe, it } from "vitest";
 // @ts-expect-error: no types
 import htmlPlugin from "textlint-plugin-html";
-import noTodoRule from "./fixtures/rule/no-todo.js";
+import noTodoRule from "./fixtures/rule/no-todo";
 // @ts-expect-error: no types
 import maxNumberOfLineRule from "textlint-rule-max-number-of-lines";
-import { TextLintTester } from "../src/textlint-tester.js";
+import { TextLintTester } from "../lib/src/textlint-tester.js";
 
 const tester = new TextLintTester();
 const baseCase = {
