+++ packages/textlint-tester/test/textlint-tester-test.ts
@@ -1,6 +1,6 @@
 import * as path from "node:path";
-import TextLintTester from "../src/index.js";
-import noTodo from "./fixtures/rule/no-todo.js";
+import TextLintTester from "../lib/src/index.js";
+import noTodo from "./fixtures/rule/no-todo";
 // @ts-expect-error: no types
 import maxNumberOfLine from "textlint-rule-max-number-of-lines";
